Refactor the advocate_interim_claim API namespace
